add generated graphql queries back into source control
These are actually supposed to stay in source control because you
literally can't regenerate them from scratch otherwise.

The reason is the following circular dep: You need the server running to
generate the queries, but you need the queries generated for the server
to start because you probably are using the queries somewhere in your
code.

If you're quick enough, you can still start the nextjs server and
immediately run the query generation command. That's BS though.